libcamera: pipeline_handler: Prepare Request
Before queueing a request to the device, any synchronization fence from
the Request framebuffers has to be waited on.

Connect the Request::Private::prepared signal to the function that
queues requests to the hardware and call Request::Private::prepare().

When the waiting request queue is inspected, verify if it has completed its
preparation phase and queue it to the device.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>