libcamera: Restrict .gitignore build/ and patches/ to the root directory
The build/ and patches/ directories are listed in .gitignore as they're
commonly used respectively as the meson build directory (this is the
directory name used in README.rst) and as a local patches store (this
usage pattern isn't official though, and may need to be reconsidered if
not used anymore). As those two directories are supposed to be at the
root of the source tree, restrict the corresponding .gitignore patterns
to avoid matching other locations.

While at it, sort the entries alphabetically, with the absolute path
entries separate from the entries that match any location within the
source tree.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo@jmondi.org>