test: fence: Signal fence once
The unit test associates a fence with a framebuffer and makes sure
the fence is correctly signalled.

Once a fence is correctly signalled, it is released by the core, and the
underlying file descriptor closed.

The unit test however tries to write on the fence file descriptor every
time the designated Request is queued, an error which is now visible
as the return value of the fence signalling write() is checked.

Fix that by associating a fence with a framebuffer only once.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Tested-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>